requirements.txt
streamlit==1.29.0
pypdf2==3.0.1
stripe==7.8.2
datetime==5.3
tenacity==8.2.3
docx2txt==0.8
textract==1.6.5
numpy==1.26.2
langchain==0.0.349
azure.identity==1.15.0
azure.keyvault.secrets==4.7.0
validate_email==1.3
openai==1.3.8
scikit-learn==1.3.2
tiktoken==0.5.2
requests==2.31.0
youtube-transcript-api==0.6.1
yt-dlp==2023.11.16
pydub==0.25.1
pytube==15.0.0
unstructured==0.11.2
openpyxl==3.1.2
pandas==2.1.4
pydantic==2.5.2
altair==5.2.0
python-dotenv==1.0.0
langchain_experimental==0.0.46